requirements.txt
langchain_openai
langchain_community
langchain_text_splitters
langchain_core
langchain_classic
langchain_pinecone
streamlit
python-dotenv
ollama
langchain-ollama